fix: remove enableAsColl param field from AaveV3SupplyAction
When supplying asset to AaveV3, asset will be enabled as collateral by
default if possible. Also, but removing this logic, transaction will not fail
when using onBehalfOfAddress. Previous, when onBehalfOfAddress was
used, call to setUserUseReserveAsCollateral would fail because msg.sender
is smart wallet which does not have any aTokens, because they are
transfered to onbehalfOf address.